#41111 Move some files into more descriptive locations
