+++ GenderPayGap.WebUI/BackgroundJobs/ScheduledJobs/FetchCompaniesHouseDataJob.cs
@@ -7,7 +7,7 @@ using GenderPayGap.Database;
 using GenderPayGap.Extensions;
 using GenderPayGap.Extensions.AspNetCore;
 
-namespace GenderPayGap.WebUI.BackgroundJobs.Jobs
+namespace GenderPayGap.WebUI.BackgroundJobs.ScheduledJobs
 {
     public class FetchCompaniesHouseDataJob
     {
