+++ GenderPayGap.WebUI/BackgroundJobs/ScheduledJobs/PurgeOrganisationsJob.cs
@@ -12,7 +12,7 @@ using GenderPayGap.Extensions.AspNetCore;
 using Microsoft.EntityFrameworkCore;
 using Newtonsoft.Json;
 
-namespace GenderPayGap.WebUI.BackgroundJobs.Jobs
+namespace GenderPayGap.WebUI.BackgroundJobs.ScheduledJobs
 {
     public class PurgeOrganisationsJob
     {
